Update to SurlyDatabase
Since LinkedLists do not have a search function, I implemented a binary search [O(log n) time] to retrieve a desired Relation object by name.

(To clarify, you can use LinkedList.indexOf(Object o), but we do not have a reference to a Relation object except for it's String name, so this is not possible.)

Added createRelation() and destroyRelation().