Filled in Relation class
Added print() (basic implementation for now, will get more specific during testing), insert(), getName(), and delete().

Note delete() is different from destroy() (which is called in SurlyDatabase). Deletion is to remove all data under the Relation, but still keep the Relation in the database. 
Destroy is to erase the Relation and its contents from the database entirely.