Changes to Attribute, AttributeValue, Tuple
Added binary search to tuple. Probably overkill but it's better than overriding the Object.equals() method. Main issue with linked lists is searching for an unknown object. Accessing by index is constant time but searching is not.
Still need to test the search functions in Tuple and SurlyDatabase.

Very minor changes to Attribute and AttributeValue classes, as they are just meant to hold data values.