Added final boolean canModify member to Relation class
The purpose is to differentiate between temporary tables that cannot have INSERT, DELETE, DESTROY executed upon them, and original, non-temporary tables that can be freely modified with those operations.

Still have to add functionality to the respective INSERT, DELETE, DESTROY methods to first verify if a Relation object is modifiable or not.